backend/requirements.txt
Flask==3.1.0
requests==2.28.1
beautifulsoup4==4.11.1
flask-cors==3.0.10
pyotp==2.9.0
logzero==1.7.0
websocket-client==1.3.3
pycryptodome==3.15.0
smartapi-python
